add CacheSkewed::replace()

+++ util/random.hpp
@@ -5,7 +5,7 @@
 
 extern void cm_set_random_seed(uint64_t seed);
 extern uint64_t cm_get_random_uint64();
-extern uint64_t cm_get_random_uint32();
+extern uint32_t cm_get_random_uint32();
 
 #include "cryptlib.h"
 #include "tiger.h"
